Fix reading Int64 values.

+++ IL Reader/MethodBodyReader.cs
@@ -177,11 +177,11 @@ namespace ILReader
         }
         private static ulong ReadInt64(byte[] il, ref int position)
         {
-            return (ulong)(il[position++] | (il[position++] << 8) | (il[position++] << 0x10) | (il[position++] << 0x18) | (il[position++] << 0x20) | (il[position++] << 0x28) | (il[position++] << 0x30) | (il[position++] << 0x38));
+            return il[position++] | ((ulong)il[position++] << 8) | ((ulong)il[position++] << 0x10) | ((ulong)il[position++] << 0x18) | ((ulong)il[position++] << 0x20) | ((ulong)il[position++] << 0x28) | ((ulong)il[position++] << 0x30) | ((ulong)il[position++] << 0x38);
         }
         private static double ReadDouble(byte[] il, ref int position)
         {
-            return il[position++] | (il[position++] << 8) | (il[position++] << 0x10) | (il[position++] << 0x18) | (il[position++] << 0x20) | (il[position++] << 0x28) | (il[position++] << 0x30) | (il[position++] << 0x38);
+            return il[position++] | ((ulong)il[position++] << 8) | ((ulong)il[position++] << 0x10) | ((ulong)il[position++] << 0x18) | ((ulong)il[position++] << 0x20) | ((ulong)il[position++] << 0x28) | ((ulong)il[position++] << 0x30) | ((ulong)il[position++] << 0x38);
         }
         private static sbyte ReadSByte(byte[] il, ref int position)
         {
